added:列表和统计

+++ app/src/main/java/com/loyo/giftapplication/room_data/salary.kt
@@ -4,4 +4,4 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class Salary(@PrimaryKey val yearAndMonth: Int, val monthSalary: Float)
+data class Salary(@PrimaryKey val yearAndMonth: Int, val monthSalary: Int)
